Update doctor and hospital handling with UI improvements
Enhanced database queries to join doctors with hospital names and provided structured error logging for SQL exceptions. Updated entity and template files to reflect these changes, added new HTML pages for hospital management, and improved UI for doctor and hospital functionalities.

+++ app/src/main/java/com/example/app/dao/HospitalRealization.java
@@ -16,6 +16,10 @@ public class HospitalRealization extends BaseDAO implements HospitalDAO {
     public void delete_hospital(int id) {
         try {
             Connection conn = getConn();
+            DoctorRealization dr = new DoctorRealization();
+            List d =dr.filterDoctors(null,String.valueOf(id),null);
+            if(!d.isEmpty())
+                return;
             String select = "DELETE FROM pz5.hospital WHERE id = ?;";
             PreparedStatement ps = getConn().prepareStatement(select);
             ps.setInt(1,id);
